feat: add function and cronjob CRUD

+++ packages/app-api/src/service/DomainService.ts
@@ -76,8 +76,17 @@ export class DomainService extends NOT_DOMAIN_SCOPED_TakaroService<DomainModel> 
   }
 
   async removeDomain(id: string) {
+    const existing = await this.findOne(id);
+
+    if (!existing) {
+      return;
+    }
+
     const gameServerService = new GameServerService(id);
-    await gameServerService.manager.remove(id);
+    const allGameServers = await gameServerService.find({});
+    for (const gameServer of allGameServers) {
+      await gameServerService.manager.remove(gameServer.id);
+    }
     await this.repo.delete(id);
   }
